change Control::{set,get}_float to Control::{set,get}_double and make almost all Control-related stuff use double rather than random mixtures of float, double and some integer value ; make hans' program change stuff work quite a bit better (not finished yet) including keyboard edting of pgm change values
git-svn-id: svn://localhost/ardour2/branches/3.0@7514 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/ardour_ui.h
@@ -368,8 +368,8 @@ class ARDOUR_UI : public Gtkmm2ext::UI, public ARDOUR::SessionHandlePtr
 	    };
 
 	    TransportControllable (std::string name, ARDOUR_UI&, ToggleType);
-	    void set_value (float);
-	    float get_value (void) const;
+	    void set_value (double);
+	    double get_value (void) const;
 
 	    void set_id (const std::string&);
