+++ libs/ardour/monitor_processor.cc
@@ -18,7 +18,7 @@ using namespace std;
 /* specialize for bool because of set_value() semantics */
 
 namespace ARDOUR {
-        template<> void MPControl<bool>::set_value (float v) {
+        template<> void MPControl<bool>::set_value (double v) {
                 bool newval = fabs (v) >= 0.5;
                 if (newval != _value) {
                         _value = newval;
